requirements.txt
bresenham==0.2
cycler==0.10.0
graphviz==0.8.2
kiwisolver==1.0.1
matplotlib==2.2.0
neat-python==0.92
numpy==1.14.1
Pillow==5.0.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
six==1.11.0
